agentcore-cdk/agent_container/requirements.txt
strands-agents
strands-agents-tools
uv
boto3
bedrock-agentcore
requests